#29 Refactored the use of GradientButton to MainActionButton.
